fix - redo GA
