+++ app/[locale]/layout.tsx
@@ -4,9 +4,9 @@ import "./globals.css";
 import Script from "next/script";
 import { usePathname } from "next/navigation";
 import { useEffect } from "react";
-
 import { NextIntlClientProvider } from "next-intl";
 import { getMessages } from "next-intl/server";
+import GoogleAnalytics from "@/components/GoogleAnalytics";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -23,41 +23,14 @@ export default async function RootLayout({
   params: { locale: string };
 }) {
   const messages = await getMessages();
-  const pathname = usePathname();
-  const GA_TRACKING_ID = process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS;
-
-  useEffect(() => {
-    if (typeof window !== "undefined" && window.gtag) {
-      window.gtag("config", GA_TRACKING_ID, {
-        page_path: pathname,
-      });
-    }
-  }, [pathname, GA_TRACKING_ID]);
 
   return (
     <html lang={locale}>
       <head>
         <link rel="icon" href="/logo.png" />
-        {GA_TRACKING_ID && (
-          <>
-            <Script
-              strategy="afterInteractive"
-              src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`}
-            />
-            <Script id="google-analytics" strategy="afterInteractive">
-              {`
-                window.dataLayer = window.dataLayer || [];
-                function gtag(){dataLayer.push(arguments);}
-                gtag('js', new Date());
-                gtag('config', '${GA_TRACKING_ID}', {
-                  page_path: window.location.pathname,
-                });
-              `}
-            </Script>
-          </>
-        )}
       </head>
       <body className={inter.className}>
+        <GoogleAnalytics />
         <NextIntlClientProvider messages={messages}>
           {children}
         </NextIntlClientProvider>
