+++ components/GoogleAnalytics.tsx
@@ -1,0 +1,42 @@
+"use client";
+
+import { useEffect } from "react";
+import Script from "next/script";
+import { usePathname } from "next/navigation";
+
+const GA_TRACKING_ID = process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS;
+
+export default function GoogleAnalytics() {
+  const pathname = usePathname();
+
+  useEffect(() => {
+    if (typeof window !== "undefined" && window.gtag) {
+      window.gtag("config", GA_TRACKING_ID, {
+        page_path: pathname,
+      });
+    }
+  }, [pathname]);
+
+  return (
+    <>
+      {GA_TRACKING_ID && (
+        <>
+          <Script
+            strategy="afterInteractive"
+            src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`}
+          />
+          <Script id="google-analytics" strategy="afterInteractive">
+            {`
+              window.dataLayer = window.dataLayer || [];
+              function gtag(){dataLayer.push(arguments);}
+              gtag('js', new Date());
+              gtag('config', '${GA_TRACKING_ID}', {
+                page_path: window.location.pathname,
+              });
+            `}
+          </Script>
+        </>
+      )}
+    </>
+  );
+}
